Move email checks and validation methods to concern
